Implement changes required by our code reviewer
