+++ lib/players.rb
@@ -1,0 +1,10 @@
+class Players
+  attr_reader :match_players
+
+  def initialize(name1, name2)
+    @match_players = [
+      { name: name1, symbol: 'X' },
+      { name: name2, symbol: 'O' }
+    ]
+  end
+end
